Handles URL destinations.
