+++ RevenueCatUI/Templates/Components/Button/ButtonComponentView.swift
@@ -13,12 +13,15 @@
 
 import Foundation
 import RevenueCat
+import SafariServices
 import SwiftUI
 
 #if PAYWALL_COMPONENTS
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 struct ButtonComponentView: View {
+    @Environment(\.openURL) var openURL
+    @State private var inAppBrowserURL: URL?
 
     private let viewModel: ButtonComponentViewModel
 
@@ -28,11 +31,72 @@ struct ButtonComponentView: View {
 
     var body: some View {
         Button(
-            action: { viewModel.onClick() },
+            action: { onClick() },
             label: { StackComponentView(viewModel: viewModel.stackViewModel) }
-        )
+        ).sheet(isPresented: .isNotNil($inAppBrowserURL)) {
+            SafariView(url: inAppBrowserURL!)
+        }
     }
 
+    private func onClick() {
+        switch viewModel.component.action {
+        case .restorePurchases:
+            // swiftlint:disable:next todo
+            // TODO handle restoring purchases
+            break
+        case .navigateTo(let destination):
+            navigateTo(destination: destination)
+        case .navigateBack:
+            // swiftlint:disable:next todo
+            // TODO handle navigating back
+            break
+        }
+    }
+
+    private func navigateTo(destination: PaywallComponent.ButtonComponent.Destination) {
+        switch destination {
+        case .customerCenter:
+            // swiftlint:disable:next todo
+            // TODO handle navigating to customer center
+            break
+        case .URL(let urlLid, let method),
+                .privacyPolicy(let urlLid, let method),
+                .terms(let urlLid, let method):
+            navigateToUrl(urlLid: urlLid, method: method)
+        }
+    }
+
+    private func navigateToUrl(urlLid: String, method: PaywallComponent.ButtonComponent.URLMethod) {
+        guard let urlString = try? viewModel.localizedStrings.string(key: urlLid),
+        let url = URL(string: urlString) else {
+            Logger.error(Strings.paywall_invalid_url(urlLid))
+            return
+        }
+
+        switch method {
+        case .inAppBrowser:
+            inAppBrowserURL = url
+        case .externalBrowser,
+                .deepLink:
+            openURL(url)
+        }
+    }
+
+}
+
+private struct SafariView: UIViewControllerRepresentable {
+    let url: URL
+
+    func makeUIViewController(context: UIViewControllerRepresentableContext<SafariView>) -> SFSafariViewController {
+        SFSafariViewController(url: url)
+    }
+
+    func updateUIViewController(
+        _ uiViewController: SFSafariViewController,
+        context: UIViewControllerRepresentableContext<SafariView>
+    ) {
+        // No updates needed
+    }
 }
 
 #if DEBUG
