+++ RevenueCatUI/Templates/Components/Button/ButtonComponentViewModel.swift
@@ -21,8 +21,8 @@ import RevenueCat
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 public class ButtonComponentViewModel {
 
-    private let component: PaywallComponent.ButtonComponent
-    private let localizedStrings: PaywallComponent.LocalizationDictionary
+    internal let component: PaywallComponent.ButtonComponent
+    internal let localizedStrings: PaywallComponent.LocalizationDictionary
     let stackViewModel: StackComponentViewModel
 
     init(
@@ -39,54 +39,6 @@ public class ButtonComponentViewModel {
             localizedStrings: localizedStrings,
             offering: offering
         )
-    }
-
-    func onClick() {
-        switch component.action {
-        case .restorePurchases:
-            // swiftlint:disable:next todo
-            // TODO handle restoring purchases
-            break
-        case .navigateTo(let destination):
-            navigateTo(destination: destination)
-        case .navigateBack:
-            // swiftlint:disable:next todo
-            // TODO handle navigating back
-            break
-        }
-    }
-
-    private func navigateTo(destination: PaywallComponent.ButtonComponent.Destination) {
-        switch destination {
-        case .customerCenter:
-            // swiftlint:disable:next todo
-            // TODO handle navigating to customer center
-            break
-        case .URL(let urlLid, let method),
-                .privacyPolicy(let urlLid, let method),
-                .terms(let urlLid, let method):
-            navigateToUrl(urlLid: urlLid, method: method)
-        }
-    }
-
-    private func navigateToUrl(urlLid: String, method: PaywallComponent.ButtonComponent.URLMethod) {
-        guard let urlString = try? localizedStrings.string(key: urlLid),
-        let url = URL(string: urlString) else {
-            Logger.error(Strings.paywall_invalid_url(urlLid))
-            return
-        }
-
-        switch method {
-        case .inAppBrowser:
-            // swiftlint:disable:next todo
-            // TODO handle navigating to URL
-            break
-        case .externalBrowser,
-                .deepLink:
-            // swiftlint:disable:next todo
-            // TODO handle navigating to URL
-            break
-        }
     }
 
 }
